move utils and types into their own folders
