+++ src/components/Chart/DataTable.tsx
@@ -1,4 +1,4 @@
-import { Props } from './utils/schema';
+import { Props } from '@/types/schema';
 
 const DataTable: React.FC<{ renderData: Props[] }> = ({ renderData }) => {
     return (
